fix: artifact "crystall ball" action, update dialog quick info
git-svn-id: https://fheroes2.svn.sourceforge.net/svnroot/fheroes2/trunk@2207 4f904287-c015-0410-838d-95d3a2656228

+++ fheroes2/src/fheroes2/heroes/heroes.h
@@ -216,7 +216,7 @@ public:
     void Scoute(void);
     u8   GetScoute(void) const;
     bool CanScouteTile(s32) const;
-    u8   GetVisionsDistance(void) const;
+    u16  GetVisionsDistance(void) const;
 
     bool isShipMaster(void) const;
     bool CanPassToShipMaster(const Heroes &) const;
